requirements.txt
amqp==5.3.1
asgiref==3.8.1
async-timeout==5.0.1
backports.zoneinfo==0.2.1
billiard==4.2.1
celery==5.3.0
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
dj-database-url==0.5.0
Django==4.2
django-cors-headers==4.4.0
django-redis==5.4.0
django-rest-framework==0.1.0
djangorestframework==3.15.0
gunicorn==21.2.0
kombu==5.5.4
packaging==25.0
prompt_toolkit==3.0.51
psycopg2-binary==2.9
python-dateutil==2.9.0.post0
python-decouple==3.8
redis==4.6.0
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.13.2
tzdata==2025.2
vine==5.1.0
wcwidth==0.2.13
whitenoise==6.7.0
